feat: add support for optional env var replacements in .npmrc (#8359)
<!-- What / Why -->
This solves problem described in #8335 in a backwards-compatible way.

<!-- Describe the request in detail. What it does and why it's being
changed. -->
This PR adds possibility to have env var replacements in .npmrc configs
written as `${VAR?}` which will cause them to get replaced with an empty
string if the variable is not defined. Old behavior where undefined
variables are left unreplaced is not changed.

## References
<!-- Examples:
  Related to #0
  Depends on #0
  Blocked by #0
  Fixes #0
  Closes #0
-->
Fixes #8335

---------

Co-authored-by: Michael Smith <owlstronaut@github.com>